refactor: 카페 상세정보 조회 도메인으로 변경
