+++ src/main/java/com/example/demo/cafe/presentation/CafeController.java
@@ -1,5 +1,9 @@
 package com.example.demo.cafe.presentation;
 
+import com.example.demo.cafe.domain.BusinessHour;
+import com.example.demo.cafe.domain.Cafe;
+import com.example.demo.cafe.service.BusinessHourService;
+import com.example.demo.cafe.service.CafeDetailResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,10 +14,6 @@ import com.example.demo.cafe.service.CafeQueryService;
 import com.example.demo.util.TimeUtil;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.servlet.mvc.Controller;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Constructor;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,11 +21,16 @@ import java.lang.reflect.Constructor;
 public class CafeController {
 
 	private final CafeQueryService cafeQueryService;
+	private final BusinessHourService businessHourService;
 	private final TimeUtil timeUtil;
 
 	@GetMapping("/{cafeId}")
 	public ResponseEntity<CafeDetailResponse> getCafeDetail(@PathVariable Long cafeId) {
-		CafeDetailResponse response = cafeQueryService.getCafeDetail(cafeId, timeUtil.now());
+		Cafe cafe = cafeQueryService.getCafe(cafeId);
+		BusinessHour businessHour = businessHourService.findBusinessHour(cafeId, timeUtil.now());
+		boolean isOpen = businessHourService.isOpen(businessHour, timeUtil.now());
+
+		CafeDetailResponse response = CafeDetailResponse.of(cafe, businessHour, isOpen);
 		return ResponseEntity.ok(response);
 	}
 }
