--- src/main/java/com/example/demo/study/domain/SearchCriteria.java
@@ -1,16 +1,0 @@
-package com.example.demo.study.domain;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDate;
-
-@Getter
-@Builder
-public class SearchCriteria {
-
-    private String keyword;
-    private LocalDate date;
-    private CafeStudyTagType cafeStudyTagType;
-    private MemberComms memberComms;
-}
